requirements.txt
# Data manipulation & analysis
pandas==2.1.4
numpy==1.26.4

# PostgreSQL connectivity
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25

# Fake data generation
Faker==22.6.0

# Configuration parsing
PyYAML==6.0.1
python-dotenv==1.0.0

# Logging & utilities
loguru==0.7.2

# Testing & coverage
pytest==7.4.4
pytest-cov==4.1.0

